tests/test_dataset.py
import pytest
from datetime import datetime
from generic_exporters.dataset import Dataset
from generic_exporters.timeseries import TimeSeries

class TestDataset:
    def test_dataset_initialization(self):
        data = TimeSeries(None, datetime.utcnow(), None, None)  # Placeholder for actual TimeSeries data
        dataset = Dataset(data)
        assert dataset._data == data

    # Placeholder tests for plot, to_csv, export methods if implemented
    def test_plot(self):
        pass

    def test_to_csv(self):
        pass

    def test_export(self):
        pass